Se corrige alumno y se agregan los mappers necesarios

+++ src/main/java/com/grup14/luterano/repository/TutorRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface TutorRepository extends JpaRepository<Tutor,Integer> {
+public interface TutorRepository extends JpaRepository<Tutor,Long> {
 
     Optional<Tutor> findByNombre (String nombre);
 }
